App done

+++ server/src/controllers/categories.controller.ts
@@ -7,8 +7,7 @@ import {CreateCategoryBody , UpdateCategoryBody }from '../types'
 export const getAllCategories = async (req: Request, res: Response) => {
     try{
 
-    const user_id = 1
-    const result = await query(`SELECT id , name , type FROM categories WHERE user_id = $1`, [user_id])
+    const result = await query(`SELECT id , name , type FROM categories WHERE user_id = $1`, [req.user!.id])
     return res.json(result.rows)
 
     } catch(error){
@@ -22,9 +21,8 @@ export const getAllCategories = async (req: Request, res: Response) => {
 
 export const createCategory = async (req : Request , res:Response) => {
     try{
-    const user_id = 1
     const {name , type} : CreateCategoryBody = req.body
-    const result = await query (`INSERT INTO categories (user_id, name, type) VALUES($1, $2, $3 ) RETURNING id , name , type`, [user_id, name, type])
+    const result = await query (`INSERT INTO categories (user_id, name, type) VALUES($1, $2, $3 ) RETURNING id , name , type`, [req.user!.id, name, type])
     return res.json(result.rows[0])
     } catch(error) {
         console.log('Error creating category:', error)
@@ -35,10 +33,9 @@ export const createCategory = async (req : Request , res:Response) => {
 
 export const updateCategory = async (req : Request , res:Response) => {
     try{
-    const user_id = 1
     const {category_id} = req.params
     const {name , type} : UpdateCategoryBody = req.body
-    await query (`UPDATE categories SET name = COALESCE($1,name), type = COALESCE($2,type) WHERE user_id = $3 AND id = $4`, [name, type, user_id, category_id])
+    await query (`UPDATE categories SET name = COALESCE($1,name), type = COALESCE($2,type) WHERE user_id = $3 AND id = $4`, [name, type, req.user!.id, category_id])
     return res.json({message : 'Successfully updated category'})
     } catch (error){
         console.log('Error could not update category:',error)
@@ -50,7 +47,7 @@ export const updateCategory = async (req : Request , res:Response) => {
 export const deleteCategory = async (req : Request, res : Response) => {
     try{
     const {category_id} = req.params
-    await query(`DELETE FROM categories WHERE id = $1`, [category_id])
+    await query(`DELETE FROM categories WHERE id = $1 AND user_id = $2`, [category_id, req.user!.id])
     res.json({ message: 'Category deleted successfully' })
     }catch(error){
         console.log('Error could not delete category:', error)
